update json images clean up dashboard components
