+++ frontend/src/components/PreviewCard/styles.css
@@ -1,0 +1,38 @@
+.preview-item-card {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
+.preview-item-card > img {
+    height: 70%;
+    width: 100%;
+    object-fit: cover;
+}
+.preview-card-description {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    position: relative;
+    background-color: var(--secondary-color);
+    height: 30%;
+    padding: 1rem;
+    margin: 0;
+}
+.preview-card-description > h3 {
+    margin: 0;
+    width: calc(100% - 1rem);
+    overflow: hidden;
+}
+.preview-card-description > img {
+    position: absolute;
+    right: 1rem;
+    height: 1rem;
+    width: 1rem;
+}
+.preview-card-description > p {
+    flex: 1;
+    margin: 0.5rem 0;
+    overflow: hidden;
+}
