+++ frontend/src/pages/Dashboard/Dashboard.css
@@ -9,6 +9,7 @@ ul {
 section {
     margin-top: 5rem;
 }
+
 .dir-btn {
     position: absolute;
     align-self: center;
@@ -37,90 +38,9 @@ section {
     text-shadow: 0 0px 1px hsla(0, 0%, 0%, 0.6), 0 1px 2px hsla(0, 0%, 0%, 0.2);
 }
 
-
-/* Similar items section  */
-#similar-items-wrapper {
-    position: relative;
-    box-sizing: content-box;
-    left: -3rem;
+#dashboard {
     width: 100%;
-    background-color: var(--accent-color);
     padding-left: 3rem;
     padding-right: 3rem;
-    min-height: 30rem;
-    max-height: 40rem;
-    height: 50vw;
-}
-#similar-items-section > h1 {
-    text-align: center;
-    display: block;
-}
-#similar-items-section {
-    box-sizing: border-box;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    padding-bottom: 2rem;
-}
-#similar-items-section > ul {
-    flex: 1;
-    display: flex;
-    flex-direction: row;
-    gap: 1rem;
-    overflow-y: hidden;
-}
-#similar-items-section > ul > li {
-    box-sizing: border-box;
-    flex: 1;
-    height: 100%;
-    background-color: white;
-}
-.similar-item-card {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-.similar-item-card > img {
-    height: 70%;
-    width: 100%;
-    object-fit: cover;
-}
-.similar-card-description {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    position: relative;
-    background-color: var(--secondary-color);
-    height: 30%;
-    padding: 1rem;
-    margin: 0;
-}
-.similar-card-description > h3 {
-    margin: 0;
-    width: calc(100% - 1rem);
-    overflow: hidden;
-}
-.similar-card-description > img {
-    position: absolute;
-    right: 1rem;
-    height: 1rem;
-    width: 1rem;
-}
-.similar-card-description > p {
-    flex: 1;
-    margin: 0.5rem 0;
-    overflow: hidden;
-}
-.hor-description {
-    height: 100%;
-}
-
-
-
-
-/* Search section  */
-#search-section {
-    text-align: center;
+    min-width: 60rem;
 }
