+++ frontend/src/pages/Dashboard/SearchSection/styles.css
@@ -1,0 +1,4 @@
+/* Search section  */
+#search-section {
+    text-align: center;
+}
