+++ frontend/src/pages/Dashboard/SimilarItemsSection/styles.css
@@ -35,41 +35,3 @@
     height: 100%;
     background-color: white;
 }
-.similar-item-card {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-.similar-item-card > img {
-    height: 70%;
-    width: 100%;
-    object-fit: cover;
-}
-.similar-card-description {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    position: relative;
-    background-color: var(--secondary-color);
-    height: 30%;
-    padding: 1rem;
-    margin: 0;
-}
-.similar-card-description > h3 {
-    margin: 0;
-    width: calc(100% - 1rem);
-    overflow: hidden;
-}
-.similar-card-description > img {
-    position: absolute;
-    right: 1rem;
-    height: 1rem;
-    width: 1rem;
-}
-.similar-card-description > p {
-    flex: 1;
-    margin: 0.5rem 0;
-    overflow: hidden;
-}
